Break out geometry read/write logic
