+++ gdal/ogr/ogrsf_frmts/flatgeobuf/GNUmakefile
@@ -2,7 +2,7 @@
 
 include ../../../GDALmake.opt
 
-OBJ	=  ogrflatgeobufdataset.o ogrflatgeobuflayer.o packedrtree.o
+OBJ	=  geometryread.o geometrywrite.o ogrflatgeobufdataset.o ogrflatgeobuflayer.o packedrtree.o
 
 CPPFLAGS	:=	-I. -I.. -I../..  $(CPPFLAGS)
 
@@ -11,5 +11,5 @@ default:	$(O_OBJ:.o=.$(OBJ_EXT))
 clean:
 	rm -f $(O_OBJ) *.o *.lo $(O_OBJ:.o=.$(OBJ_EXT))
 
-$(O_OBJ):	ogr_flatgeobuf.h
+$(O_OBJ):	ogr_flatgeobuf.h cplerrors.h
